+++ conn/client_job.go
@@ -4,7 +4,12 @@ import (
 	"github.com/hashicorp/nomad/api"
 )
 
-// Run is parsing the HCL code and registering the Job into Nomad
-func (co *Client) GetJob(code string) (*api.Job, error) {
-	return co.c.Jobs().ParseHCL(code, true)
+// GetJobFromCode is parsing the HCL code and providing a api.Job{} pointer
+func (co *Client) GetJobFromCode(code string) (*api.Job, error) {
+	return co.jobs.ParseHCL(code, true)
+}
+
+func (co *Client) GetJobStatus(jobId string) (j *api.Job, err error) {
+	j, _, err = co.jobs.Info(jobId, nil)
+	return
 }